mktables: Fix up version compare
This is a feature that is used to compare 2 different Unicode versions
for changes in existing code points as opposed to additional code points
or properties.  It does this by removing the newly-added code points
coming from the later version.  One can then diff the directory
structure generated with the older version to see what changed.

Prior to this commit, it assumed all version numbers were single digit
for the major number.  This no longer works for Unicode 10.

This commit also causes mktables to more completely add blocks that
didn't exist in the earlier version back to the unallocated pool.  This
gives better diff results.