XXX Use first part of this for merge commit msg pp_ctl.c: Replace loop by using memchr
I did an audit of core for loops that exit when they have found a
particular character.  All such loops have an upper limit that they are
not to look beyond, which may not be explicit.  I changed these loops
to instead use memchr().  This is cleaner, and may be faster.

Some loops look for a particular character, but don't exit when they
find one; instead they do special processing, and keep going until the
end of the input is reached.  In these cases, I retained the loop, but
changed to use memchr() to find the character of interest.  Again, this
may be faster.

I made a separate commit in this branch for each changed loop.  This is
so, should I have messed up, a bisect will immediately finger the exact
culprit.

See the merge commit message for more information.