utf8.c: Don't calc code point from overflowing UTF8
This avoids calculating a code point from UTF-8 that is known to
overflow.  This could give incorrect results (used only in warning
messages), but is done only when there are 3 (or more) malformations:
overflow, overlong, UTF-8 terminated early, so it's unlikely to actually
happen in the field.

I am not adding any tests, as I don't know of any existing failures, and
making code points be at most IV_MAX would cause existing tests to fail
without having to add new ones.  I imagine a brute force generator of
UTF-8 would find some string that showed this problem up absent the
other coming changes, but it's not worth it.