APItest/t/utf8_setup.pl: Add #define equivalent
I don't know of an easy way to automatically import hdr file constants
into Perl code (and if there were one, I there are plenty of existing
modules that don't take advantage of it), and so they get copied and
pasted into Perl, and changed enough to match Perl syntax.  When the
constants change, the Perl code must be manually updated.  Here, a new
constant was added, and now the Perl is being updated to match.